+++ app/src/main/java/com/jeremyhahn/cropdroid/model/ConditionConfig.kt
@@ -1,8 +1,8 @@
 package com.jeremyhahn.cropdroid.model
 
-data class ConditionConfig(var id: Int, val metricId: Long, var channelId: Long, val comparator: String, val threshold: Double) {
+data class ConditionConfig(var id: Long, val metricId: Long, var channelId: Long, val comparator: String, val threshold: Double) {
 
-    constructor(id: Int) : this( 0, 0, 0, "", 0.0) {
+    constructor(id: Long) : this( 0, 0, 0, "", 0.0) {
         this.id = id
     }
 }
