+++ app/src/main/java/com/jeremyhahn/cropdroid/ui/condition/ConditionListRecyclerAdapter.kt
@@ -43,13 +43,13 @@ class ConditionListRecyclerAdapter(val activity: ConditionListActivity, val crop
 
             Log.d("onCreateContextMenu", "condition: " + condition)
 
-            menu!!.add(0, condition.id, 0, "Edit")
+            menu!!.add(0, condition.id.toInt(), 0, "Edit")
                 .setOnMenuItemClickListener(MenuItem.OnMenuItemClickListener() {
                     activity.showConditionDialog(condition)
                     true
                 })
 
-            menu!!.add(0, condition.id, 0, "Delete")
+            menu!!.add(0, condition.id.toInt(), 0, "Delete")
                 .setOnMenuItemClickListener(MenuItem.OnMenuItemClickListener() {
                     activity.deleteCondition(ConditionConfig(condition.id))
                     true
